+++ src/test/java/io/github/astrapi69/color/CssColorToJavaColorExtensions.java
@@ -1,3 +1,27 @@
+/**
+ * The MIT License
+ *
+ * Copyright (C) 2021 Asterios Raptis
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
 package io.github.astrapi69.color;
 
 import java.awt.Color;
@@ -5,8 +29,13 @@ import java.util.regex.Pattern;
 
 import io.github.astrapi69.color.model.HsbModel;
 
+/**
+ * The class {@link CssColorToJavaColorExtensions} provides view extension methods for
+ * transformation of css color strings to {@link Color} objects and back
+ */
 public class CssColorToJavaColorExtensions
 {
+
 	private static final Pattern hexRegex = Pattern.compile("#[\\dA-Fa-f]{6}");
 	private static final Pattern hexRegexWithTransparency = Pattern.compile("#[\\dA-Fa-f]{8}");
 	private static final Pattern rgbRegex = Pattern.compile("rgba?\\([^)]*\\)",