+++ src/test/java/io/github/astrapi69/color/CssColorToJavaColorExtensionsTest.java
@@ -1,15 +1,65 @@
+/**
+ * The MIT License
+ *
+ * Copyright (C) 2021 Asterios Raptis
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
 package io.github.astrapi69.color;
 
 import static org.junit.jupiter.api.Assertions.*;
 
 import java.awt.Color;
+import java.util.List;
 
 import org.junit.jupiter.api.Test;
+import org.meanbean.test.BeanTester;
 
-import io.github.astrapi69.color.model.HsbModel;
+import io.github.astrapi69.collection.list.ListFactory;
 
+
+/**
+ * Test class for the class {@link CssColorToJavaColorExtensions}
+ */
 class CssColorToJavaColorExtensionsTest
 {
+
+	/**
+	 * Test method for {@link CssColorToJavaColorExtensions#toColor(String)}
+	 */
+	@Test
+	void toColor()
+	{
+		CssColorToJavaColorExtensionsTest.getTestCssStrings().stream().forEach(
+			cssColor -> System.out.println(CssColorToJavaColorExtensions.toColor(cssColor)));
+	}
+
+	/**
+	 * Test data
+	 */
+	public static List<String> getTestCssStrings()
+	{
+		List<String> cssColorStrings = ListFactory.newArrayList("#ff00cc", "#FF00CC", "#ff000080",
+			"#FF00CC80", "rgb(255,0,0)", "rgba(255,0,0,0.5)", "rgba(100%,0%,30%,0.5)");
+		return cssColorStrings;
+	}
 
 	/**
 	 * Test method for {@link CssColorToJavaColorExtensions#toHexString(Color, boolean)}
@@ -189,12 +239,14 @@ class CssColorToJavaColorExtensionsTest
 
 	}
 
-	Color toColor(int i)
+	/**
+	 * Test method for {@link CssColorToJavaColorExtensions} with {@link BeanTester}
+	 */
+	@Test
+	public void testWithBeanTester()
 	{
-		final int a = (i >> 24) & 0xFF;
-		final int r = (i >> 16) & 0xFF;
-		final int g = (i >> 8) & 0xFF;
-		final int b = i & 0xFF;
-		return new Color(r, g, b, a);
+		final BeanTester beanTester = new BeanTester();
+		beanTester.testBean(CssColorToJavaColorExtensions.class);
 	}
+
 }